// src/components/TradingViewWidgets/types.tsx
export type CopyrightStyles = {
    parent?: React.CSSProperties;
    link?: React.CSSProperties;
    span?: React.CSSProperties;
};